requirements.txt
python-telegram-bot
requests~=2.28.1
machine==0.0.1